build(plugins): introduce rspack as replacement for webpack
