+++ plugins/dbgate-plugin-cassandra/package.json
@@ -15,9 +15,9 @@
     "icon.svg"
   ],
   "scripts": {
-    "build:frontend": "webpack --config webpack-frontend.config",
-    "build:frontend:watch": "webpack  --watch --config webpack-frontend.config",
-    "build:backend": "webpack --config webpack-backend.config.js",
+    "build:frontend": "rspack --config rspack-frontend.config.js",
+    "build:frontend:watch": "rspack  --watch --config rspack-frontend.config.js",
+    "build:backend": "rspack --config rspack-backend.config.js",
     "build": "yarn build:frontend && yarn build:backend",
     "plugin": "yarn build && yarn pack && dbgate-plugin dbgate-plugin-cassandra",
     "plugout": "dbgate-plugout dbgate-plugin-cassandra",
@@ -25,14 +25,14 @@
     "prepublishOnly": "yarn build"
   },
   "devDependencies": {
-    "dbgate-plugin-tools": "^1.0.8",
-    "webpack": "^5.91.0",
-    "webpack-cli": "^5.1.4"
+    "@rspack/cli": "^1.3.11",
+    "@rspack/core": "^1.3.11",
+    "dbgate-plugin-tools": "^1.0.8"
   },
   "dependencies": {
+    "cassandra-driver": "^4.7.2",
     "dbgate-tools": "^6.0.0-alpha.1",
     "json-stable-stringify": "^1.0.1",
-    "lodash": "^4.17.21",
-    "cassandra-driver": "^4.7.2"
+    "lodash": "^4.17.21"
   }
 }
