+++ plugins/dbgate-plugin-cassandra/rspack-backend.config.js
@@ -1,4 +1,4 @@
-var webpack = require('webpack');
+const { rspack } = require('@rspack/core');
 var path = require('path');
 
 const packageJson = require('./package.json');
