+++ plugins/dbgate-plugin-dbf/package.json
@@ -22,9 +22,9 @@
     "icon.svg"
   ],
   "scripts": {
-    "build:frontend": "webpack --config webpack-frontend.config",
-    "build:frontend:watch": "webpack --watch --config webpack-frontend.config",
-    "build:backend": "webpack --config webpack-backend.config.js",
+    "build:frontend": "rspack --config rspack-frontend.config.js",
+    "build:frontend:watch": "rspack --watch --config rspack-frontend.config.js",
+    "build:backend": "rspack --config rspack-backend.config.js",
     "build": "yarn build:frontend && yarn build:backend",
     "plugin": "yarn build && yarn pack && dbgate-plugin dbgate-plugin-dbf",
     "copydist": "yarn build && yarn pack && dbgate-copydist ../dist/dbgate-plugin-dbf",
@@ -32,9 +32,9 @@
     "prepublishOnly": "yarn build"
   },
   "devDependencies": {
-    "dbgate-plugin-tools": "^1.0.7",
-    "webpack": "^5.91.0",
-    "webpack-cli": "^5.1.4"
+    "@rspack/cli": "^1.3.11",
+    "@rspack/core": "^1.3.11",
+    "dbgate-plugin-tools": "^1.0.7"
   },
   "dependencies": {
     "dbffile": "^1.12.0",
