+++ plugins/dbgate-plugin-mssql/package.json
@@ -21,9 +21,9 @@
     "icon.svg"
   ],
   "scripts": {
-    "build:frontend": "webpack --config webpack-frontend.config",
-    "build:frontend:watch": "webpack --watch --config webpack-frontend.config",
-    "build:backend": "webpack --config webpack-backend.config.js",
+    "build:frontend": "rspack --config rspack-frontend.config.js",
+    "build:frontend:watch": "rspack --watch --config rspack-frontend.config.js",
+    "build:backend": "rspack --config rspack-backend.config.js",
     "build": "yarn build:frontend && yarn build:backend",
     "prepublishOnly": "yarn build",
     "plugin": "yarn build && yarn pack && dbgate-plugin dbgate-plugin-mssql",
@@ -31,9 +31,9 @@
     "plugout": "dbgate-plugout dbgate-plugin-mssql"
   },
   "devDependencies": {
-    "dbgate-plugin-tools": "^1.0.7",
-    "webpack": "^5.91.0",
-    "webpack-cli": "^5.1.4"
+    "@rspack/cli": "^1.3.11",
+    "@rspack/core": "^1.3.11",
+    "dbgate-plugin-tools": "^1.0.7"
   },
   "dependencies": {
     "@azure/identity": "^4.6.0",
