+++ plugins/dbgate-plugin-mysql/rspack-frontend.config.js
@@ -1,4 +1,5 @@
-var webpack = require('webpack');
+const { rspack } = require('@rspack/core');
+
 var path = require('path');
 
 var config = {
@@ -8,15 +9,22 @@ var config = {
     app: './index.js',
   },
   target: 'web',
+  experiments: {
+    outputModule: true,
+  },
   output: {
     path: path.resolve(__dirname, 'dist'),
     filename: 'frontend.js',
-    libraryTarget: 'var',
-    library: 'plugin',
+    library: {
+      name: 'plugin',
+      type: 'var',
+      export: 'default',
+    },
+    module: true,
   },
 
   plugins: [
-    new webpack.DefinePlugin({
+    new rspack.DefinePlugin({
       'global.DBGATE_PACKAGES': 'window.DBGATE_PACKAGES',
     }),
   ],
