+++ plugins/dbgate-plugin-oracle/package.json
@@ -20,9 +20,9 @@
     "icon.svg"
   ],
   "scripts": {
-    "build:frontend": "webpack --config webpack-frontend.config",
-    "build:frontend:watch": "webpack --watch --config webpack-frontend.config",
-    "build:backend": "webpack --config webpack-backend.config.js",
+    "build:frontend": "rspack --config rspack-frontend.config.js",
+    "build:frontend:watch": "rspack --watch --config rspack-frontend.config.js",
+    "build:backend": "rspack --config rspack-backend.config.js",
     "build": "yarn build:frontend && yarn build:backend",
     "plugin": "yarn build && yarn pack && dbgate-plugin dbgate-plugin-oracle",
     "copydist": "yarn build && yarn pack && dbgate-copydist ../dist/dbgate-plugin-oracle",
@@ -30,9 +30,9 @@
     "prepublishOnly": "yarn build"
   },
   "devDependencies": {
-    "dbgate-plugin-tools": "^1.0.8",
-    "webpack": "^5.91.0",
-    "webpack-cli": "^5.1.4"
+    "@rspack/cli": "^1.3.11",
+    "@rspack/core": "^1.3.11",
+    "dbgate-plugin-tools": "^1.0.8"
   },
   "dependencies": {
     "dbgate-query-splitter": "^4.11.5",
