+++ plugins/dbgate-plugin-redis/package.json
@@ -19,9 +19,9 @@
     "dist"
   ],
   "scripts": {
-    "build:frontend": "webpack --config webpack-frontend.config",
-    "build:frontend:watch": "webpack --watch --config webpack-frontend.config",
-    "build:backend": "webpack --config webpack-backend.config.js",
+    "build:frontend": "rspack --config rspack-frontend.config.js",
+    "build:frontend:watch": "rspack --watch --config rspack-frontend.config.js",
+    "build:backend": "rspack --config rspack-backend.config.js",
     "build": "yarn build:frontend && yarn build:backend",
     "plugin": "yarn build && yarn pack && dbgate-plugin dbgate-plugin-redis",
     "plugout": "dbgate-plugout dbgate-plugin-redis",
@@ -29,16 +29,16 @@
     "prepublishOnly": "yarn build"
   },
   "devDependencies": {
-    "dbgate-plugin-tools": "^1.0.7",
-    "webpack": "^5.91.0",
-    "webpack-cli": "^5.1.4"
+    "@rspack/cli": "^1.3.11",
+    "@rspack/core": "^1.3.11",
+    "dbgate-plugin-tools": "^1.0.7"
   },
   "dependencies": {
+    "async": "^3.2.3",
     "dbgate-query-splitter": "^4.11.5",
     "dbgate-tools": "^6.0.0-alpha.1",
-    "lodash": "^4.17.21",
-    "async": "^3.2.3",
     "ioredis": "^5.4.1",
+    "lodash": "^4.17.21",
     "node-redis-dump2": "^0.5.0"
   }
 }
