+++ plugins/dbgate-plugin-sqlite/package.json
@@ -20,9 +20,9 @@
     "icon.svg"
   ],
   "scripts": {
-    "build:frontend": "webpack --config webpack-frontend.config",
-    "build:frontend:watch": "webpack --watch --config webpack-frontend.config",
-    "build:backend": "webpack --config webpack-backend.config.js",
+    "build:frontend": "rspack --config rspack-frontend.config.js",
+    "build:frontend:watch": "rspack --watch --config rspack-frontend.config.js",
+    "build:backend": "rspack --config rspack-backend.config.js",
     "build": "yarn build:frontend && yarn build:backend",
     "plugin": "yarn build && yarn pack && dbgate-plugin dbgate-plugin-sqlite",
     "copydist": "yarn build && yarn pack && dbgate-copydist ../dist/dbgate-plugin-sqlite",
@@ -30,17 +30,17 @@
     "prepublishOnly": "yarn build"
   },
   "devDependencies": {
-    "dbgate-plugin-tools": "^1.0.4",
-    "webpack": "^5.91.0",
-    "webpack-cli": "^5.1.4"
+    "@rspack/cli": "^1.3.11",
+    "@rspack/core": "^1.3.11",
+    "dbgate-plugin-tools": "^1.0.4"
   },
   "dependencies": {
+    "dbgate-query-splitter": "^4.11.5",
     "dbgate-tools": "^6.0.0-alpha.1",
-    "lodash": "^4.17.21",
-    "dbgate-query-splitter": "^4.11.5"
+    "lodash": "^4.17.21"
   },
   "optionalDependencies": {
-    "libsql": "0.5.0-pre.6",
-    "better-sqlite3": "11.8.1"
+    "better-sqlite3": "11.8.1",
+    "libsql": "0.5.0-pre.6"
   }
 }
